// services/core-api/src/models/GameGenre.ts
export enum GameGenre {
    ALL = 0,
    POP = 132,
    ROCK = 152,
    METAL = 464,
    RAP = 116,
    RNB = 165,
    FOLK = 466,
    COUNTRY = 84,
    FRENCH = 52,
    SOUL = 169,
    BLUES = 153
}

export const GameGenreValues = [
    GameGenre.ALL,
    GameGenre.POP,
    GameGenre.ROCK,
    GameGenre.METAL,
    GameGenre.RAP,
    GameGenre.RNB,
    GameGenre.FOLK,
    GameGenre.COUNTRY,
    GameGenre.FRENCH,
    GameGenre.SOUL,
    GameGenre.BLUES,
]